Cleanup for RecoverJobsTest
Issue#93
